+++ primitives/mod.ts
@@ -1,3 +1,2 @@
-export * from "./Era.ts"
 export * from "./MultiAddress.ts"
 export * from "./MultiSignature.ts"
