+++ reflection/mod.ts
@@ -1,3 +1,4 @@
 export * from "./Codec.ts"
+export * from "./redirects/mod.ts"
 export * from "./Ty.ts"
 export * from "./TyVisitor.ts"
